+++ NativeCollections/NativeCollections/NativeConcurrentDictionary.cs
@@ -101,8 +101,8 @@ namespace NativeCollections
                 locks[i] = new object();
             var countPerLock = new NativeArray<int>(locks.Length, true);
             var buckets = new NativeArray<VolatileNode>(capacity, true);
-            _handle = (NativeConcurrentDictionaryHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeConcurrentDictionaryHandle));
-            _handle->Tables = (Tables*)NativeMemoryAllocator.Alloc(sizeof(Tables));
+            _handle = (NativeConcurrentDictionaryHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeConcurrentDictionaryHandle));
+            _handle->Tables = (Tables*)NativeMemoryAllocator.Alloc((uint)sizeof(Tables));
             _handle->Tables->Initialize(buckets, locks, countPerLock);
             _handle->GrowLockArray = growLockArray;
             _handle->Budget = buckets.Length / locks.Length;
@@ -549,7 +549,7 @@ namespace NativeCollections
 
                 var newBuckets = new NativeArray<VolatileNode>(newLength, true);
                 var newCountPerLock = new NativeArray<int>(newLocks.Length, true);
-                var newTables = (Tables*)NativeMemoryAllocator.Alloc(sizeof(Tables));
+                var newTables = (Tables*)NativeMemoryAllocator.Alloc((uint)sizeof(Tables));
                 newTables->Initialize(newBuckets, newLocks, newCountPerLock);
                 AcquirePostFirstLock(tables, ref locksAcquired);
                 foreach (var bucket in tables->Buckets)
