+++ NativeCollections/NativeCollections/NativeList.cs
@@ -61,8 +61,8 @@ namespace NativeCollections
                 throw new ArgumentOutOfRangeException(nameof(capacity), capacity, "MustBeNonNegative");
             if (capacity < 4)
                 capacity = 4;
-            _handle = (NativeListHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeListHandle));
-            _handle->Array = (T*)NativeMemoryAllocator.Alloc(capacity * sizeof(T));
+            _handle = (NativeListHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeListHandle));
+            _handle->Array = (T*)NativeMemoryAllocator.Alloc((uint)(capacity * sizeof(T)));
             _handle->Length = capacity;
             _handle->Size = 0;
             _handle->Version = 0;
@@ -119,7 +119,7 @@ namespace NativeCollections
                 {
                     if (value > 0)
                     {
-                        var newItems = (T*)NativeMemoryAllocator.Alloc(value * sizeof(T));
+                        var newItems = (T*)NativeMemoryAllocator.Alloc((uint)(value * sizeof(T)));
                         if (_handle->Size > 0)
                             Unsafe.CopyBlockUnaligned(newItems, _handle->Array, (uint)(_handle->Size * sizeof(T)));
                         NativeMemoryAllocator.Free(_handle->Array);
