+++ NativeCollections/NativeCollections/NativeMemoryPool.cs
@@ -124,7 +124,7 @@ namespace NativeCollections
             if (maxFreeSlabs < 0)
                 throw new ArgumentOutOfRangeException(nameof(maxFreeSlabs), maxFreeSlabs, "MustBeNonNegative");
             var nodeSize = sizeof(NativeMemoryNode) + length;
-            var array = (byte*)NativeMemoryAllocator.Alloc(sizeof(NativeMemorySlab) + size * nodeSize);
+            var array = (byte*)NativeMemoryAllocator.Alloc((uint)(sizeof(NativeMemorySlab) + size * nodeSize));
             var slab = (NativeMemorySlab*)array;
             slab->Next = slab;
             slab->Previous = slab;
@@ -139,7 +139,7 @@ namespace NativeCollections
 
             slab->Node = next;
             slab->Count = size;
-            _handle = (NativeMemoryPoolHandle*)NativeMemoryAllocator.Alloc(sizeof(NativeMemoryPoolHandle));
+            _handle = (NativeMemoryPoolHandle*)NativeMemoryAllocator.Alloc((uint)sizeof(NativeMemoryPoolHandle));
             _handle->Slab = slab;
             _handle->FreeSlab = null;
             _handle->Slabs = 1;
@@ -269,7 +269,7 @@ namespace NativeCollections
                     if (_handle->FreeSlabs == 0)
                     {
                         var nodeSize = sizeof(NativeMemoryNode) + _handle->Length;
-                        var array = (byte*)NativeMemoryAllocator.Alloc(sizeof(NativeMemorySlab) + size * nodeSize);
+                        var array = (byte*)NativeMemoryAllocator.Alloc((uint)(sizeof(NativeMemorySlab) + size * nodeSize));
                         slab = (NativeMemorySlab*)array;
                         array += sizeof(NativeMemorySlab);
                         NativeMemoryNode* next = null;
